internal/contour: move http conn manager filter to internal/envoy
Updates #537
Updates #708

Move contour.httpfilter to envoy.HTTPConnectionManager.
