+++ internal/contour/listener.go
@@ -21,10 +21,20 @@ import (
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/listener"
 	"github.com/gogo/protobuf/proto"
-	"github.com/gogo/protobuf/types"
 	"github.com/heptio/contour/internal/dag"
 	"github.com/heptio/contour/internal/envoy"
 	"k8s.io/api/core/v1"
+)
+
+const (
+	ENVOY_HTTP_LISTENER            = "ingress_http"
+	ENVOY_HTTPS_LISTENER           = "ingress_https"
+	DEFAULT_HTTP_ACCESS_LOG        = "/dev/stdout"
+	DEFAULT_HTTP_LISTENER_ADDRESS  = "0.0.0.0"
+	DEFAULT_HTTP_LISTENER_PORT     = 8080
+	DEFAULT_HTTPS_ACCESS_LOG       = "/dev/stdout"
+	DEFAULT_HTTPS_LISTENER_ADDRESS = DEFAULT_HTTP_LISTENER_ADDRESS
+	DEFAULT_HTTPS_LISTENER_PORT    = 8443
 )
 
 // ListenerCache manages the contents of the gRPC LDS cache.
@@ -174,23 +184,6 @@ func (c *listenerCache) Values(filter func(string) bool) []proto.Message {
 	return values
 }
 
-const (
-	ENVOY_HTTP_LISTENER            = "ingress_http"
-	ENVOY_HTTPS_LISTENER           = "ingress_https"
-	DEFAULT_HTTP_ACCESS_LOG        = "/dev/stdout"
-	DEFAULT_HTTP_LISTENER_ADDRESS  = "0.0.0.0"
-	DEFAULT_HTTP_LISTENER_PORT     = 8080
-	DEFAULT_HTTPS_ACCESS_LOG       = "/dev/stdout"
-	DEFAULT_HTTPS_LISTENER_ADDRESS = DEFAULT_HTTP_LISTENER_ADDRESS
-	DEFAULT_HTTPS_LISTENER_PORT    = 8443
-
-	router     = "envoy.router"
-	grpcWeb    = "envoy.grpc_web"
-	gzip       = "envoy.gzip"
-	httpFilter = "envoy.http_connection_manager"
-	accessLog  = "envoy.file_access_log"
-)
-
 type listenerVisitor struct {
 	*ListenerCache
 	dag.Visitable
@@ -207,7 +200,7 @@ func (v *listenerVisitor) Visit() map[string]*v2.Listener {
 		},
 	}
 	filters := []listener.Filter{
-		httpfilter(ENVOY_HTTPS_LISTENER, v.httpsAccessLog()),
+		envoy.HTTPConnectionManager(ENVOY_HTTPS_LISTENER, v.httpsAccessLog()),
 	}
 	v.Visitable.Visit(func(vh dag.Vertex) {
 		switch vh := vh.(type) {
@@ -240,7 +233,7 @@ func (v *listenerVisitor) Visit() map[string]*v2.Listener {
 			Name:    ENVOY_HTTP_LISTENER,
 			Address: socketaddress(v.httpAddress(), v.httpPort()),
 			FilterChains: []listener.FilterChain{
-				filterchain(v.UseProxyProto, httpfilter(ENVOY_HTTP_LISTENER, v.httpAccessLog())),
+				filterchain(v.UseProxyProto, envoy.HTTPConnectionManager(ENVOY_HTTP_LISTENER, v.httpAccessLog())),
 			},
 		}
 	}
@@ -274,45 +267,6 @@ func filterchain(useproxy bool, filters ...listener.Filter) listener.FilterChain
 	return fc
 }
 
-func httpfilter(routename, accessLogPath string) listener.Filter {
-	return listener.Filter{
-		Name: httpFilter,
-		Config: &types.Struct{
-			Fields: map[string]*types.Value{
-				"stat_prefix": sv(routename),
-				"rds": st(map[string]*types.Value{
-					"route_config_name": sv(routename),
-					"config_source": st(map[string]*types.Value{
-						"api_config_source": st(map[string]*types.Value{
-							"api_type": sv("GRPC"),
-							"grpc_services": lv(
-								st(map[string]*types.Value{
-									"envoy_grpc": st(map[string]*types.Value{
-										"cluster_name": sv("contour"),
-									}),
-								}),
-							),
-						}),
-					}),
-				}),
-				"http_filters": lv(
-					st(map[string]*types.Value{
-						"name": sv(gzip),
-					}),
-					st(map[string]*types.Value{
-						"name": sv(grpcWeb),
-					}),
-					st(map[string]*types.Value{
-						"name": sv(router),
-					}),
-				),
-				"use_remote_address": {Kind: &types.Value_BoolValue{BoolValue: true}}, // TODO(jbeda) should this ever be false?
-				"access_log":         accesslog(accessLogPath),
-			},
-		},
-	}
-}
-
 func tlscontext(data map[string][]byte, tlsMinProtoVersion auth.TlsParameters_TlsProtocol, alpnprotos ...string) *auth.DownstreamTlsContext {
 	return &auth.DownstreamTlsContext{
 		CommonTlsContext: &auth.CommonTlsContext{
@@ -335,26 +289,4 @@ func tlscontext(data map[string][]byte, tlsMinProtoVersion auth.TlsParameters_Tl
 			AlpnProtocols: alpnprotos,
 		},
 	}
-}
-
-func accesslog(path string) *types.Value {
-	return lv(
-		st(map[string]*types.Value{
-			"name": sv(accessLog),
-			"config": st(map[string]*types.Value{
-				"path": sv(path),
-			}),
-		}),
-	)
-}
-
-func sv(s string) *types.Value {
-	return &types.Value{Kind: &types.Value_StringValue{StringValue: s}}
-}
-
-func st(m map[string]*types.Value) *types.Value {
-	return &types.Value{Kind: &types.Value_StructValue{StructValue: &types.Struct{Fields: m}}}
-}
-func lv(v ...*types.Value) *types.Value {
-	return &types.Value{Kind: &types.Value_ListValue{ListValue: &types.ListValue{Values: v}}}
 }
