+++ internal/envoy/listener.go
@@ -25,3 +25,67 @@ func TLSInspector() listener.ListenerFilter {
 		Config: new(types.Struct),
 	}
 }
+
+// HTTPConnectionManager creates a new HTTP Connection Manager filter
+// for the supplied route and access log.
+func HTTPConnectionManager(routename, accessLogPath string) listener.Filter {
+	return listener.Filter{
+		Name: "envoy.http_connection_manager",
+		Config: &types.Struct{
+			Fields: map[string]*types.Value{
+				"stat_prefix": sv(routename),
+				"rds": st(map[string]*types.Value{
+					"route_config_name": sv(routename),
+					"config_source": st(map[string]*types.Value{
+						"api_config_source": st(map[string]*types.Value{
+							"api_type": sv("GRPC"),
+							"grpc_services": lv(
+								st(map[string]*types.Value{
+									"envoy_grpc": st(map[string]*types.Value{
+										"cluster_name": sv("contour"),
+									}),
+								}),
+							),
+						}),
+					}),
+				}),
+				"http_filters": lv(
+					st(map[string]*types.Value{
+						"name": sv("envoy.gzip"),
+					}),
+					st(map[string]*types.Value{
+						"name": sv("envoy.grpc_web"),
+					}),
+					st(map[string]*types.Value{
+						"name": sv("envoy.router"),
+					}),
+				),
+				"use_remote_address": {Kind: &types.Value_BoolValue{BoolValue: true}}, // TODO(jbeda) should this ever be false?
+				"access_log":         accesslog(accessLogPath),
+			},
+		},
+	}
+}
+
+func accesslog(path string) *types.Value {
+	return lv(
+		st(map[string]*types.Value{
+			"name": sv("envoy.file_access_log"),
+			"config": st(map[string]*types.Value{
+				"path": sv(path),
+			}),
+		}),
+	)
+}
+
+func sv(s string) *types.Value {
+	return &types.Value{Kind: &types.Value_StringValue{StringValue: s}}
+}
+
+func st(m map[string]*types.Value) *types.Value {
+	return &types.Value{Kind: &types.Value_StructValue{StructValue: &types.Struct{Fields: m}}}
+}
+
+func lv(v ...*types.Value) *types.Value {
+	return &types.Value{Kind: &types.Value_ListValue{ListValue: &types.ListValue{Values: v}}}
+}
